fix(SetModel.java): switch weight to double

+++ src/main/java/com/gymtracker/data/model/SetModel.java
@@ -19,7 +19,7 @@ public class SetModel {
   public Integer repetitions;
 
   @Column(name = "weight", nullable = false)
-  public Integer weight;
+  public Double weight;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "workout_exercise_id", nullable = false)
